Add Emu.from_key_or_nil
`from_key_or_nil` behaves the same as `Hash#[key]` in that it returns
`nil` if the key doesn't exist in the map. An alternative and probably
nicer design would have been `from_key? :: Key -> Decoder a -> Decoder
(Maybe a)`. Ruby not shipping with a `Maybe` type doesn't make this very
practical, though. I had essentially the following choices:

1. Define a custom `Maybe` type myself. This makes it very hard to
   compose the results of Emu with other libraries working with a
   different `Maybe`.
2. Use one of the existing libraries. `dry-monads` looks to be the most
   used one of all of them, but unfortunately their `Maybe#fmap` is broken
   and last time I raised that issue it seemed like that behavior is
   intended:

    Some(5).fmap { |_| nil } # => None

   This leaves me with one of the other libraries, but none of them seem
   to be used. Choosing a less used library would have had the same
   effect as going with option 1, plus I'd have to maintain a
   dependency. So far, Emu has none.